First private combat (no lasers drawn, though). Many changes in tactics.s and tinefuncs.s
In tactics.s there is a routine called AIMain where all the AI is implemented (equivalently to tactics in tactics.c in AGB sources). This is called just for objects with the IS_AICONTROLLED flag set in ai_state flags. The approach_tactics function in AGB is here called fly_to_ship. Splitted fly_to_vector in get_attack_angle and fly_to_vector_final and touched others to conform this.

The second ship (ID3) goes towards the first one (ID2) and shoots, that makes the first get angry and they engage in combat.