Corrected a bug where thargons were not given the ECM. Also thargoids carried are random (0-3).
Tested that it is possible to fight a thargon. It is, if well equipped, but the trick of firing a missile when very close is quite helpful.

Adapted the ship size to the scaled-down versions.

Corrected the bug where planet_distance was not correctly set at the first frame. Now you can't jump just after undocking.

Corrected the visibility distance, so it is no more possible that a ship in front of you fires, while it is not visible (lasers incorrectly drawn).

Added shipdatal.xls (with the scaling factor and size) to the repository.