More bug hunting and small additions. The most important thing is that I have added Dbug's new line routine, with no extra pixel error.
Now if ECM is active enemies don't launch missiles. Also find target makes sure that a target has not been deleted even if still in the visible list (could happen). This and other checks seem to minimize the error where invisible strange enemies were killed and gave you a lot of credits. But the problem is still there. I have set a halt if it happens for more testing.