Small minor additions and some cleaning of code.
Added a worm and a missile counter, so we can limit their number for computer AIs. Limits in defines in tine.h

Radar and scanner are updated only every 2 frames.

Asteroids now rotate.

Found a bug in the radar. It needs a lot of space to work with all the possible objects, but that is indeed a waste, so need to clean up that code, which is indeed a mess.

This version has ALTSCANS undefined for testing purposes.