A quick backup. Several changes, small details all. Included the new version of the LineDraw.s, also made some small optimizations to lib3d.s.
Moved commander's data to start of page 4 to ease mission code, but it is not definite.

Also changed the way bounty messages are issued, so they appear with some kind of delay and accumulate bounties till then.

Added a patch to erase laser vertex only of ship objects. Should use this to limit the number of space objects.