Work in progress on a makefile engine for OSDK. Loosely inspired by the BeOS makefile engine.
This allows using a Makefile template with a few lines to fill out instead of unportable batch files.
It also eases upgrading the engine since it is not copied within the project directories.

Currently the NMAKE-based engine fully compiles things and runs the selected emulator.
When the NAME is not set in the makefile an include file is generated from osdk_config.bat to help migrating.
The GNU make version is TODO.